Fix #8: unify yaks and yaks-cli and add release scripts
