+++ .gitignore
@@ -7,6 +7,6 @@ build/_maven_dependencies
 build/_output
 
 /yaks
-/yaks-cli
+/yaks-*
 
 target/
