+++ pkg/cmd/root.go
@@ -22,7 +22,6 @@ import (
 	"os"
 
 	"github.com/jboss-fuse/yaks/pkg/client"
-	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
 )
 
@@ -43,10 +42,9 @@ func NewYaksCommand(ctx context.Context) (*cobra.Command, error) {
 		Context: ctx,
 	}
 	var cmd = cobra.Command{
-		PersistentPreRunE: options.preRun,
-		Use:               "yaks",
-		Short:             "Yaks is a awesome client tool for running tests natively on Kubernetes",
-		Long:              yaksCommandLongDescription,
+		Use:   "yaks",
+		Short: "Yaks is a awesome client tool for running tests natively on Kubernetes",
+		Long:  yaksCommandLongDescription,
 	}
 
 	cmd.PersistentFlags().StringVar(&options.KubeConfig, "config", os.Getenv("KUBECONFIG"), "Path to the config file to use for CLI requests")
@@ -54,36 +52,7 @@ func NewYaksCommand(ctx context.Context) (*cobra.Command, error) {
 
 	cmd.AddCommand(newCmdTest(&options))
 	cmd.AddCommand(newCmdInstall(&options))
+	cmd.AddCommand(newCmdOperator(&options))
 
 	return &cmd, nil
-}
-
-func (command *RootCmdOptions) preRun(cmd *cobra.Command, _ []string) error {
-	if command.Namespace == "" {
-		current, err := client.GetCurrentNamespace(command.KubeConfig)
-		if err != nil {
-			return errors.Wrap(err, "cannot get current namespace")
-		}
-		err = cmd.Flag("namespace").Value.Set(current)
-		if err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-// GetCmdClient returns the client that can be used from command line tools
-func (command *RootCmdOptions) GetCmdClient() (client.Client, error) {
-	// Get the pre-computed client
-	if command._client != nil {
-		return command._client, nil
-	}
-	var err error
-	command._client, err = command.NewCmdClient()
-	return command._client, err
-}
-
-// NewCmdClient returns a new client that can be used from command line tools
-func (command *RootCmdOptions) NewCmdClient() (client.Client, error) {
-	return client.NewOutOfClusterClient(command.KubeConfig)
 }
