+++ pkg/cmd/util.go
@@ -1,0 +1,38 @@
+package cmd
+
+import (
+	"github.com/jboss-fuse/yaks/pkg/client"
+
+	"github.com/pkg/errors"
+	"github.com/spf13/cobra"
+)
+
+func (command *RootCmdOptions) preRun(cmd *cobra.Command, _ []string) error {
+	if command.Namespace == "" {
+		current, err := client.GetCurrentNamespace(command.KubeConfig)
+		if err != nil {
+			return errors.Wrap(err, "cannot get current namespace")
+		}
+		err = cmd.Flag("namespace").Value.Set(current)
+		if err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
+// GetCmdClient returns the client that can be used from command line tools
+func (command *RootCmdOptions) GetCmdClient() (client.Client, error) {
+	// Get the pre-computed client
+	if command._client != nil {
+		return command._client, nil
+	}
+	var err error
+	command._client, err = command.NewCmdClient()
+	return command._client, err
+}
+
+// NewCmdClient returns a new client that can be used from command line tools
+func (command *RootCmdOptions) NewCmdClient() (client.Client, error) {
+	return client.NewOutOfClusterClient(command.KubeConfig)
+}
